move DriverResponse class outside from the models module
