+++ cloudshell/cp/core/request_actions/__init__.py
@@ -1,6 +1,7 @@
 from .cleanup_sandbox_infra import CleanupSandboxInfraRequestActions  # noqa
 from .deploy_vm import DeployVMRequestActions  # noqa
 from .deployed_app import DeployedApp  # noqa
+from .driver_response import DriverResponse  # noqa
 from .get_vm_details import GetVMDetailsRequestActions  # noqa
 from .prepare_sandbox_infra import PrepareSandboxInfraRequestActions  # noqa
 from .set_app_security_groups import SetAppSecurityGroupsRequestActions  # noqa
