+++ cloudshell/cp/core/request_actions/models/create_keys.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-from cloudshell.cp.core.requests.models import BaseRequestAction
+from .base import BaseRequestAction
 
 
 @dataclass
