+++ cloudshell/cp/core/request_actions/models/prepare_cloud_infra.py
@@ -1,6 +1,7 @@
 from dataclasses import dataclass
 
-from cloudshell.cp.core.requests.models import BaseConnectivityAction, BaseRequestObject
+from .base import BaseRequestObject
+from .connectivity import BaseConnectivityAction
 
 
 @dataclass
