+++ cloudshell/cp/core/request_actions/models/prepare_subnet.py
@@ -1,6 +1,7 @@
 from dataclasses import dataclass, field
 
-from cloudshell.cp.core.requests.models import BaseConnectivityAction, BaseRequestObject
+from .base import BaseRequestObject
+from .connectivity import BaseConnectivityAction
 
 
 @dataclass
