+++ cloudshell/cp/core/request_actions/set_app_security_groups.py
@@ -1,7 +1,6 @@
-import json
 from dataclasses import dataclass
 
-from cloudshell.cp.core.requests.actions.base import BaseRequestActions
+from cloudshell.cp.core.request_actions.base import BaseRequestActions
 
 
 @dataclass
